Move libarchive-related sources into the clp namespace.
