+++ components/core/src/clp/LibarchiveFileReader.cpp
@@ -2,8 +2,9 @@
 
 #include <cstring>
 
-#include "spdlog_with_specializations.hpp"
+#include "../spdlog_with_specializations.hpp"
 
+namespace clp {
 ErrorCode LibarchiveFileReader::try_get_pos(size_t& pos) {
     if (nullptr == m_archive) {
         throw OperationFailed(ErrorCode_NotInit, __FILENAME__, __LINE__);
@@ -268,3 +269,4 @@ ErrorCode LibarchiveFileReader::read_next_data_block() {
 
     return ErrorCode_Success;
 }
+}  // namespace clp
